Update eventDebugId type to uint16_t
- Change the `eventDebugId` global variable from `int` to `uint16_t` to ensure type consistency and prevent potential sign-related issues.

+++ src/driver/main.cpp
@@ -62,7 +62,7 @@ LDR 3 - night sensor:
 // -------------------------------
 // Global variables and constants
 // -------------------------------
-static int eventDebugId = 1;
+static uint16_t eventDebugId = 1;
 
 // ---------------------------------------------
 // Settings structures and constants
